backend/insy7315-warehouse-api/requirements.txt
# Core web framework
fastapi==0.120.4
uvicorn[standard]==0.38.0

# Database
pyodbc==5.2.0

# Environment variables
python-dotenv==1.2.1

# Password hashing
passlib[argon2]==1.7.4

# JWT handling
PyJWT==2.10.1